continuing fix for #116
