+++ modules/cmpapi/src/command/GetTCDataCommand.ts
@@ -1,46 +1,13 @@
-import {CmpApiModel} from '../CmpApiModel';
 import {Command} from './Command';
-import {GVL} from '@iabtcf/core';
 import {TCDataCallback} from '../callback';
 import {TCData} from '../response';
 
 export class GetTCDataCommand extends Command {
 
-  private isSuccess: boolean;
-
-  /**
-   * because the success is an override method and we want to perform some
-   * async action, this method is created to proxy that one.
-   */
-  private async sendData(): Promise<void> {
-
-    const tcModel = CmpApiModel.tcModel;
-
-    /**
-     * If the tcModel exists as a TCModel and doesn't have a gvl or a tcString,
-     * then we'll need a gvl to create a tcString to return in the TCData
-     * object. Otherwise gdpr does not apply and they'll get a sparse object or
-     * gdpr does apply but a tcString is set and there is no need to have GVL
-     * because the existing tcString can be used.
-     */
-
-    if (CmpApiModel.tcModel !== null && !tcModel.gvl && !CmpApiModel.tcString) {
-
-      const consentLanguage = tcModel.consentLanguage;
-      tcModel.gvl = new GVL(tcModel.vendorListVersion);
-      await tcModel.gvl.readyPromise;
-      tcModel.consentLanguage = consentLanguage;
-
-    }
+  protected success(): void {
 
     const callback = this.callback as TCDataCallback;
     callback(new TCData(this.param, this.listenerId), true);
-
-  }
-
-  protected success(): void {
-
-    this.sendData();
 
   }
