+++ modules/cmpapi/test/PageHandler.test.ts
@@ -1,0 +1,13 @@
+import {PageHandler} from '../src/PageHandler';
+import {expect} from 'chai';
+
+describe('PageHandler', (): void => {
+
+  it('\x1b[36mNeeds Unit Tests\x1b[0m', (done: () => void): void => {
+
+    expect(true).to.be.true;
+    done();
+
+  });
+
+});
