+++ modules/cmpapi/test/command/DisabledCommand.test.ts
@@ -1,0 +1,13 @@
+import {DisabledCommand} from '../../src/command/DisabledCommand';
+import {expect} from 'chai';
+
+describe('command->DisabledCommand', (): void => {
+
+  it('\x1b[36mNeeds Unit Tests\x1b[0m', (done: () => void): void => {
+
+    expect(true).to.be.true;
+    done();
+
+  });
+
+});
